Cast tv_usec to unsigned int for snprintf

+++ iso8601.c
@@ -16,6 +16,6 @@ void timestamp(char s[])
 		 * implement portable nanosecond accuracy.
 		 */
 		strftime(s, TIMESTAMP - 9, "%Y-%m-%dT%H:%M:%S.", tm);
-		snprintf(&s[TIMESTAMP - 9], 10, "%06u000", tv.tv_usec);
+		snprintf(&s[TIMESTAMP - 9], 10, "%06u000", (unsigned int)tv.tv_usec);
 	}
 }
